Cambio int a decimal

+++ TP Actual/src/PagoElectronico/Consulta Saldos/Consultar Saldo.cs
@@ -40,7 +40,7 @@ namespace PagoElectronico.Consulta_Saldos
         private void actualizarCuentas()
         {
             DataTable cuentas;
-            int user;
+            decimal user;
             
             try
             {
